+++ html/index.html
@@ -84,8 +84,9 @@
         </div>
 
         <!--<script src="../js/testangles.js" charset="utf-8"></script>-->
-        <script src="../js/objects.js"></script>
-        <script src="../js/planeControls.js" charset="utf-8"></script>
+        <script src="../js/routes.js"></script>
+        <script src="../js/planes.js" charset="utf-8"></script>
+        <script src="../js/controls.js" charset="utf-8"></script>
         <script src="../js/devTools.js" charset="utf-8"></script>
     </body>
 </html>
