+++ js/devTools.js
@@ -5,12 +5,12 @@ function testRoutes(velocity){
     var flDiff = -20
     ROUTES.forEach(function(route){
         var plane = new Plane(300+flDiff, 300+flChange, route, false, route.name, velocity*1000);
+        updateLists(plane, -1);
         createPlaneElt(plane);
         animPlane(plane);
         screenElt.appendChild(plane.elt);
         flChange +=50;
         flDiff += 10;
-
     });
 }
