--- js/points.js
@@ -1,30 +1,0 @@
-var gai = {"x":165, "y":2088};
-var etori = {"x":755, "y":1893};
-var men = {"x":774, "y":1646};
-var cfa = {"x":751, "y":866};
-var vulca = {"x":707, "y":616};
-var spidy = {"x":996, "y":1814};
-var lanza = {"x":973, "y":1457};
-var mindi = {"x":1163, "y":1277};
-var major = {"x":1540 , "y":2060};//estimated
-var mtl = {"x":1524, "y":1639};
-var ltp = {"x":1779, "y":1014};
-var lse = {"x":1615, "y":848};
-var burgo = {"x":1340, "y":383};
-var atn = {"x":1237, "y":207};
-var liman = {"x":1842, "y":661};
-var melka = {"x":2269, "y":309};
-var fri = {"x":2514, "y":106};
-var rapid = {"x":2723, "y":250};
-var sevet = {"x":2401, "y":512};
-var mozao = {"x":2086, "y":765};
-var grena = {"x":2056, "y":1463};
-var bosua = {"x":2486, "y":1321};
-var juven = {"x":2841, "y":1203};
-var santo = {"x":2265, "y":1803};
-var jambi = {"x":2500, "y":2186};
-var sicil = {"x":3030, "y":2482};
-var ouestPoint= {"x":295, "y":2045};
-var eastPoint= {"x":2764 , "y":1231};
-var = {"x": , "y": };
-var = {"x": , "y": };
